fix: adjust styles for the country popup and for the personal page
