+++ src/components/Modals/AddGroup.jsx
@@ -19,12 +19,14 @@ export const AddGroups = ({ country }) => {
         isSelected: group.regions
           .some((country1) => country1.id === country.id)
       })))
-  }, [loading])
+  }, [loading]);
+
+
   return (
     <div>
-      <h4 className={'fs-5 fw-bold'}>Add Groups</h4>
+      <h4 className={'fs-5 fw-bold'}>Add guides</h4>
       {dataWithProperties && !loading && <div>
-        <div style={{ marginTop: "2rem", display: "grid", gap: "1rem",overflowY:"auto",height:"12rem", gridTemplateColumns: "0.25fr 0.25fr 0.25fr 0.25fr" }}>
+        <div style={{ marginTop: "2rem", display: "grid", gap: "1rem",overflowY:"auto",height:"6rem", gridTemplateColumns: "0.25fr 0.25fr 0.25fr 0.25fr" }}>
           {dataWithProperties.map((group, index) => {
             return (
               <div
@@ -90,16 +92,16 @@ export const AddGroups = ({ country }) => {
                             .map((country1) => ({ id: country1.id }))]
                       }))
                 }
-              }).then((e) => {
-              message.success("Groups updated successfully");
+              }).then(() => {
+              message.success("New guide updated successfully");
               modal.reset()
             })
-              .catch((error) => {
-                message.error("Groups update failed")
+              .catch(() => {
+                message.error("Guides update failed")
               })
           }}
           style={{ cursor: "pointer", display: "flex", justifyContent: "center", alignItems: "center", position: "relative", color: "black", height: "2.3rem", minWidth: "11.5rem", background: "white", borderRadius: "0.8rem", marginRight: "1rem" }}>
-          Add groups
+          Add new guides
         </div>
       </div>
     </div>
