+++ src/views/Personalnformation/RightPersonal.jsx
@@ -4,13 +4,9 @@ import styles from "./RightPersonal.module.css";
 import useTravelRecommenderStore from "../../store/travelRecommenderStore";
 import useLoadMeWithGroups from "../../api/useLoadMeWithGroups";
 import useLoadStatistics from "../../api/useLoadStatistics";
-import useUpdateMeWithGroups from "../../api/useUpdateMeWithGroups";
 import GoToMapCountryButton from "../../components/GoToMapCountry";
 import { useState, useRef, useEffect } from "react";
-import TextWithInput from "../../components/TextWithInput";
-import { DeleteOutlined } from "@ant-design/icons";
-import { message } from "antd";
-import { EditOutlined } from "@ant-design/icons";
+import GroupRow from "../../components/GroupRow";
 
 const FavouriteRow = ({ score, region, id }) => {
   return (
@@ -19,102 +15,12 @@ const FavouriteRow = ({ score, region, id }) => {
         {score && <h6 className='fa-xs lh-1 fw-bold'>Score {Math.floor(score)}/100</h6>}
         <h5 className={`fs-6 fw-bold ${styles.groupRegions}`}>{region}</h5>
       </Col>
-      <Col className={'d-flex justify-content-evenly align-items-center ms-5'}>
+      <Col className={'d-flex justify-content-evenly align-items-center'}>
         <GoToMapCountryButton regionId={id} />
       </Col>
     </div>
   )
 };
-
-const GroupRow = ({ id, inputRef, name, regions, onCreate, setCreateNewGroup, groups, setGroups }) => {
-  const { executePutGroups } = useUpdateMeWithGroups();
-  const { fetch} = useLoadMeWithGroups();
-  const deleteGroup = () => {
-    executePutGroups({
-      data: {
-        groups: groups.filter((group) => group.id !== id)
-      }
-    }).then((response) => {
-      if (response.status < 400) {
-        setGroups(groups.filter((group) => group.id !== id))
-        message.success("Group deleted successfully")
-      } else {
-        throw new Error('Failed to delete the group')
-      }
-    }).catch((error) => {
-      message.error(error.message);
-    })
-  }
-  const onSave = (groupName) => {
-    executePutGroups({
-      data: {
-        groups: [...groups, { name: groupName, regions: [] }]
-      }
-    })
-      .then((response) => {
-        if (response.status < 400) {
-          return fetch().catch((error) => {
-            throw new Error('Failed to fetch the groups');
-          });
-        } else {
-          throw new Error(response.data.error.message || 'Failed to executePutGroups');
-        }
-      })
-      .then(() => {
-        setGroups([...groups, { name: groupName, regions: [] }]);
-        setCreateNewGroup(false);
-        message.success("Group created successfully");
-
-      })
-      .catch((error) => {
-        message.error(error.response?.data?.error?.message || error.message);
-        setCreateNewGroup(true);
-      });
-
-  }
-  return (
-    <>
-      {!onCreate && <div className={styles.elementRow}>
-        <Col className='col-8 d-flex flex-column'>
-          <div style={{ display: "flex", alignItems: "center" }}>
-            <h5 style={{ display: "inline" }} className='fs-6 fw-bold'>{name}</h5>
-            <EditOutlined
-              style={{ marginBottom: "6px", marginLeft: "1rem", display: "inline" }}
-              onClick={() => {
-              }}
-            />
-          </div>
-
-          <h6 className={`fa-xs lh-1 fw-bold ${styles.groupRegions}`}>
-            {regions.map(region => region.u_name).join(', ')}
-          </h6>
-        </Col>
-        <Col className={'d-flex justify-content-end cursor align-items-center'}>
-          <button onClick={deleteGroup} className="btn text-white">
-            <DeleteOutlined />
-          </button>
-        </Col>
-      </div>}
-
-      {onCreate &&
-        <div style={{ position: "relative", display: "block" }} className={styles.elementRow}>
-          <div style={{ marginTop: "1rem" }}>
-            <TextWithInput
-              text={""}
-              inputRef={inputRef}
-              createNewGroup={onCreate}
-              setCreateNewGroup={setCreateNewGroup}
-              style={{ marginTop: "1rem" }}
-              defaultText={'No group'}
-              onSave={onSave}
-            />
-          </div>
-        </div>
-      }
-    </>
-
-  )
-}
 
 const RightPersonal = () => {
   const { data: favourites } = useFavouritesPaginationFrontend();
@@ -190,25 +96,25 @@ const RightPersonal = () => {
       </Col>
       <Col>
         <div style={{ justifyContent: "space-between", display: "flex" }}>
-          <h5 className='fs-6 fw-bold pb-2'>Groups</h5>
+          <h5 className='fs-6 fw-bold pb-2'>Travel Guides</h5>
           <span onClick={() => {
             setNewGroupCreation(true);
           }
           }
-            style={{ cursor: "pointer", fontSize: "15px", marginRight: "25px" }}>Add group</span>
+            style={{ cursor: "pointer", fontSize: "15px", marginRight: "25px" }}>Add new guide</span>
         </div>
         <div ref={scrollRef} className={styles.groupsHeight}>
           <div className={'d-flex pe-2  flex-column gap-3'}>
-            {groups.map(group => (
+            {groups.map((group, idx) => (
               <GroupRow
-                key={group.id}
+                key={idx}
                 id={group.id}
                 name={group.name}
+                description={group.description}
                 groups={groups}
                 regions={group.regions}
                 onCreate={false}
                 setGroups={setGroups}
-
               />
             ))}
             {newGroupCreation && <div style={{ display: "block" }}>
