+++ src/views/Personalnformation/RightPersonal.module.css
@@ -4,7 +4,6 @@
     padding: 0.6rem 1.45rem;
     border-radius: 18px;
     width: 100%;
-    height: 4.6rem;
     display: flex;
     align-items: center;
     border: 1px solid #336273;
